revno: 2212 message:   Fixed issue with closed date on update orgunit
committer: Lars Helge Overland <larshelge@gmail.com>
branch nick: dhis2
timestamp: Mon 2010-11-29 15:40:20 +0100

+++ dhis-2/dhis-web/dhis-web-maintenance/dhis-web-maintenance-organisationunit/src/main/java/org/hisp/dhis/oum/action/organisationunit/GetOrganisationUnitAction.java
@@ -82,7 +82,7 @@ public class GetOrganisationUnitAction
     public String execute()
         throws Exception
     {
-        organisationUnit = organisationUnitService.getOrganisationUnit( id.intValue() );
+        organisationUnit = organisationUnitService.getOrganisationUnit( id );
 
         numberOfChildren = organisationUnit.getChildren().size();
